# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GALI \u2014 Ground-truth Analytics for Listed Issuers API",
    "description": "Production read API and live scenario simulation engine bridging upstream Indonesian mining concession intelligence (ESDM) with listed capital market equity valuations (IDX).",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/issuers": {
      "get": {
        "tags": [
          "Issuers & Fundamental Analytics"
        ],
        "summary": "List Issuers",
        "description": "List all in-universe issuers with summary ground truth scores, RLI, RBV, and data quality badges.",
        "operationId": "list_issuers_v1_issuers_get",
        "parameters": [
          {
            "name": "commodity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by commodity (e.g. Coal)",
              "title": "Commodity"
            },
            "description": "Filter by commodity (e.g. Coal)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IssuerSummary"
                  },
                  "title": "Response List Issuers V1 Issuers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/issuers/{symbol}": {
      "get": {
        "tags": [
          "Issuers & Fundamental Analytics"
        ],
        "summary": "Get Issuer Detail",
        "description": "Retrieve full fundamental intelligence report for an issuer (M1\u2013M9, linked entities, and audit evidence).",
        "operationId": "get_issuer_detail_v1_issuers__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuerDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/issuers/{symbol}/graph": {
      "get": {
        "tags": [
          "Issuers & Fundamental Analytics"
        ],
        "summary": "Get Issuer Graph",
        "description": "Generate interactive node-edge graph connecting listed issuer -> operating subsidiaries -> mining sites -> licenses -> contractors.",
        "operationId": "get_issuer_graph_v1_issuers__symbol__graph_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuerGraphResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GraphEdge": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight"
          },
          "properties": {
            "additionalProperties": true,
            "type": "object",
            "title": "Properties"
          }
        },
        "type": "object",
        "required": [
          "source",
          "target",
          "label"
        ],
        "title": "GraphEdge"
      },
      "GraphNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "type": {
            "type": "string",
            "enum": [
              "issuer",
              "operating_company",
              "mining_site",
              "license",
              "contractor"
            ],
            "title": "Type"
          },
          "properties": {
            "additionalProperties": true,
            "type": "object",
            "title": "Properties"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "type"
        ],
        "title": "GraphNode"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IssuerDetail": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "sub_sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Sector"
          },
          "as_of": {
            "type": "string",
            "format": "date",
            "title": "As Of"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "data_quality": {
            "type": "string",
            "title": "Data Quality"
          },
          "rli_years": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rli Years"
          },
          "implied_life_years": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Implied Life Years"
          },
          "reserve_life_gap_years": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reserve Life Gap Years"
          },
          "attributable_gross_profit_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attributable Gross Profit Usd"
          },
          "reserve_backed_value_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reserve Backed Value Usd"
          },
          "market_cap_idr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap Idr"
          },
          "market_cap_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap Usd"
          },
          "rbv_gap_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rbv Gap Pct"
          },
          "license_cliff_1y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Cliff 1Y"
          },
          "license_cliff_3y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Cliff 3Y"
          },
          "license_cliff_5y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Cliff 5Y"
          },
          "cnc_coverage_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cnc Coverage Pct"
          },
          "weighted_days_to_expiry": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weighted Days To Expiry"
          },
          "cash_cost_per_ton_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cash Cost Per Ton Usd"
          },
          "realized_price_per_ton_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Realized Price Per Ton Usd"
          },
          "unit_margin_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Margin Usd"
          },
          "breakeven_benchmark_price_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Breakeven Benchmark Price Usd"
          },
          "cost_curve_percentile": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Curve Percentile"
          },
          "weighted_cv_kcal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weighted Cv Kcal"
          },
          "benchmark_grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Benchmark Grade"
          },
          "benchmark_price_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Benchmark Price Usd"
          },
          "quality_discount_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Discount Pct"
          },
          "destination_hhi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Hhi"
          },
          "top_destination": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Destination"
          },
          "top_destination_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Destination Pct"
          },
          "contractor_hhi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contractor Hhi"
          },
          "contract_cliff_12m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Cliff 12M"
          },
          "ground_truth_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ground Truth Score"
          },
          "component_scores": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Component Scores"
          },
          "confidence": {
            "additionalProperties": true,
            "type": "object",
            "title": "Confidence"
          },
          "linked_entities": {
            "items": {
              "$ref": "#/components/schemas/LinkedOperatingEntity"
            },
            "type": "array",
            "title": "Linked Entities"
          },
          "evidence": {
            "additionalProperties": true,
            "type": "object",
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "name",
          "as_of",
          "run_id",
          "data_quality"
        ],
        "title": "IssuerDetail"
      },
...
      "IssuerGraphResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "type": "array",
            "title": "Edges"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "nodes",
          "edges"
        ],
        "title": "IssuerGraphResponse"
      },
...
      "IssuerSummary": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "sub_sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Sector"
          },
          "data_quality": {
            "type": "string",
            "title": "Data Quality",
            "description": "'LENGKAP' (complete) or 'PARSIAL' (partial)"
          },
          "ground_truth_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ground Truth Score",
            "description": "M8 composite score (0-100)"
          },
          "confidence_pct": {
            "type": "number",
            "title": "Confidence Pct",
            "description": "Effective weight utilized (0-100%)",
            "default": 100.0
          },
          "rli_years": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rli Years",
            "description": "M1 Reserve Life Index"
          },
          "reserve_backed_value_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reserve Backed Value Usd",
            "description": "M2 Reserve-Backed Value in USD"
          },
          "market_cap_idr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap Idr"
          },
          "market_cap_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap Usd"
          },
          "rbv_gap_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rbv Gap Pct",
            "description": "M2 valuation premia / discount %"
          },
          "license_cliff_3y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Cliff 3Y",
            "description": "M3 3-year concession expiry %"
          },
          "cash_cost_per_ton_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cash Cost Per Ton Usd",
            "description": "M4 unit cash cost FOB"
          },
          "top_destination": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Destination",
            "description": "M6 top export market"
          },
          "top_destination_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Destination Pct"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "name",
          "data_quality"
        ],
        "title": "IssuerSummary"
      },
      "LinkedOperatingEntity": {
        "properties": {
          "company_slug": {
            "type": "string",
            "title": "Company Slug"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "effective_ownership_pct": {
            "type": "number",
            "title": "Effective Ownership Pct"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method"
          }
        },
        "type": "object",
        "required": [
          "company_slug",
          "effective_ownership_pct"
        ],
        "title": "LinkedOperatingEntity"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}